Allow us-east-1 to be used in testing
us-east-1 is a special case region in S3 for historical reasons and does
not share the same convention for the base endpoint.  The details of
endpoints can be reviewed at
http://docs.aws.amazon.com/general/latest/gr/rande.html#s3_region.  This
documentation outlines a secondary way to reach all buckets that is
consistent.  Using the "dualstack" endpoints prevents us from adding
conditionals to deal with us-east-1.