Support CodeBuild/ECR
AWS CodeBuild uses ECR under the hood. Instance roles are passed in differently to containers in ECR. Instead of being available on 169.254.169.254, they are available on 169.254.170.2 in a different format. The old AWS SDK that was used doesn't know how to read the instance role from there. This commit fixes the issue by upgrading to a newer AWS SDK that is aware of the new ECR metadata server and therefore works in CodeBuild.